额外菜单优化

+++ game/Submods/NeteaseInMas/NeteaseMusicPlayerMain.rpy
@@ -642,10 +642,13 @@ init 999 python:
         persistent._NP_API_key_able = False
         store.mas_submod_utils.submod_log.info("初始化连接时发生异常：{}".format(e))
 
-init -300 python:
+init 300 python:
     try:
-        file = np_globals.Catch + "/" + ".gitkeep"
-        os.remove(file)
+        for file_name in dirs:
+            if file_name.find('git') != -1:
+                file = np_globals.Catch + "/" + file_name
+                os.remove(file)
+                break
     except:
         pass
     try:
